requirements.txt
# Core Dependencies
streamlit==1.50.0
anthropic==0.71.0
pandas==2.3.3
plotly==5.18.0
python-dotenv==1.0.0

# API Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0

# Database
psycopg[binary]

# Vector Store
pinecone
voyageai